Move NumberInput into own component and cleanup
