+++ example/src/App.tsx
@@ -55,6 +55,7 @@ import StepperExample from "./StepperExample";
 
 import TextFieldExample from "./TextFieldExample";
 import TextInputExample from "./TextInputExample";
+import NumberInputExample from "./NumberInputExample";
 import CheckboxExample from "./CheckboxExample";
 import WebViewExample from "./WebViewExample";
 import AccordionExample from "./AccordionExample";
@@ -108,6 +109,7 @@ const ROUTES = {
   Swiper: SwiperExample,
   TextField: TextFieldExample,
   TextInput: TextInputExample,
+  NumberInput: NumberInputExample,
   WebView: WebViewExample,
 };
