+++ example/src/TextInputExample.js
@@ -1,12 +1,11 @@
 import * as React from "react";
 import { TextInput, KeyboardAvoidingView } from "react-native";
-import { NumberInput, ButtonSolid } from "@draftbit/ui";
+import { ButtonSolid } from "@draftbit/ui";
 import Section, { Container, styles } from "./Section";
 
 export default function TextInputExample() {
   const [value, setText] = React.useState("Change me!");
   const [value2, setText2] = React.useState("Change me!");
-  const [number, setNumber] = React.useState(123);
   const handleChange = (text) => setText(text);
 
   return (
@@ -29,34 +28,6 @@ export default function TextInputExample() {
             value={value2}
             onChangeText={(text) => setText2(text)}
             defaultValue="I'm an initial value!"
-          />
-
-          <NumberInput
-            placeholder="Number input using value"
-            value={number}
-            onChangeText={(num) => setNumber(num)}
-          />
-          <NumberInput
-            placeholder="Number input using value (2)"
-            value={number}
-            onChangeText={(num) => setNumber(num)}
-          />
-          <NumberInput
-            placeholder="Number input using defaultValue"
-            onChangeText={(num) => setNumber(num)}
-            defaultValue={1}
-          />
-          <NumberInput
-            placeholder="Disabled number input"
-            onChangeText={(num) => setNumber(num)}
-            defaultValue={1}
-            editable={false}
-          />
-
-          <ButtonSolid
-            style={[styles.button]}
-            title="Reset Number"
-            onPress={(_) => setNumber(0)}
           />
         </Section>
       </Container>
